requirements.txt
torch
torchvision
numpy
matplotlib
streamlit
streamlit-drawable-canvas
Pillow